requirements.txt
imgurpython
Pillow
